Moved entities in their respective model directories
