+++ src/RecipeBot.Domain/Factories/RecipeDomainEntityFactory.cs
@@ -19,8 +19,8 @@ using System;
 using System.Collections.Generic;
 using Common.Utils;
 using RecipeBot.Domain.Data;
-using RecipeBot.Domain.Entities;
 using RecipeBot.Domain.Exceptions;
+using RecipeBot.Domain.Models;
 using RecipeBot.Domain.Properties;
 
 namespace RecipeBot.Domain.Factories;
