+++ src/RecipeBot.Domain/Factories/RecipeFieldDomainEntityFactory.cs
@@ -17,8 +17,8 @@
 
 using System;
 using Common.Utils;
-using RecipeBot.Domain.Entities;
 using RecipeBot.Domain.Exceptions;
+using RecipeBot.Domain.Models;
 using RecipeBot.Domain.Properties;
 
 namespace RecipeBot.Domain.Factories;
