+++ test/RecipeBot.Domain.Test/Factories/RecipeDomainEntityFactoryTest.cs
@@ -20,9 +20,9 @@ using System.Collections.Generic;
 using System.Linq;
 using NSubstitute;
 using RecipeBot.Domain.Data;
-using RecipeBot.Domain.Entities;
 using RecipeBot.Domain.Exceptions;
 using RecipeBot.Domain.Factories;
+using RecipeBot.Domain.Models;
 using RecipeBot.TestUtils;
 using Xunit;
