+++ test/RecipeBot.Domain.Test/Factories/RecipeFieldDomainEntityFactoryTest.cs
@@ -17,9 +17,9 @@
 
 using System;
 using NSubstitute;
-using RecipeBot.Domain.Entities;
 using RecipeBot.Domain.Exceptions;
 using RecipeBot.Domain.Factories;
+using RecipeBot.Domain.Models;
 using RecipeBot.TestUtils;
 using Xunit;
