+++ test/RecipeBot.Domain.Test/Models/AuthorDomainEntityTest.cs
@@ -16,58 +16,61 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 using System;
-using RecipeBot.Domain.Entities;
+using RecipeBot.Domain.Models;
 using RecipeBot.TestUtils;
 using Xunit;
 
-namespace RecipeBot.Domain.Test.Entities;
+namespace RecipeBot.Domain.Test.Models;
 
-public class RecipeFieldDomainEntityTest
+public class AuthorDomainEntityTest
 {
     [Theory]
     [ClassData(typeof(NullOrWhitespacesStringValueGenerator))]
-    public void Entity_with_invalid_field_name_throws_exception(string invalidFieldName)
+    public void Entity_with_invalid_name_throws_exception(string invalidAuthorName)
     {
         // Setup
-        const string fieldData = "fieldData";
+        const string imageUrl = "http://wwww.google.com";
 
         // Call
-        Action call = () => new RecipeFieldDomainEntity(invalidFieldName, fieldData);
+        Action call = () => new AuthorDomainEntity(invalidAuthorName, imageUrl);
 
         // Assert
         Assert.Throws<ArgumentException>(call);
     }
 
     [Theory]
-    [ClassData(typeof(NullOrWhitespacesStringValueGenerator))]
-    public void Entity_with_invalid_field_data_throws_exception(string invalidFieldData)
+    [ClassData(typeof(InvalidHttpUrlDataGenerator))]
+    [InlineData(null)]
+    [InlineData("")]
+    [InlineData("     ")]
+    public void Entity_with_invalid_image_url_throws_exception(string invalidImageUrl)
     {
         // Setup
-        const string fieldName = "fieldName";
+        const string authorName = "Author";
 
         // Call
-        Action call = () => new RecipeFieldDomainEntity(fieldName, invalidFieldData);
+        Action call = () => new AuthorDomainEntity(authorName, invalidImageUrl);
 
         // Assert
         Assert.Throws<ArgumentException>(call);
     }
 
     [Theory]
-    [InlineData("field name", "field data")]
-    [InlineData("field name", "field    data")]
-    [InlineData("field    name", "field data")]
-    [InlineData("     name", "     data")]
-    [InlineData("name     ", "data     ")]
-    public void Entity_with_valid_data_returns_total_length_of_properties(string fieldName, string fieldData)
+    [InlineData("authorName")]
+    [InlineData("author name")]
+    [InlineData("author    name")]
+    [InlineData("     authorName")]
+    [InlineData("authorName     ")]
+    public void Entity_with_valid_data_returns_total_length_of_author_name(string authorName)
     {
         // Setup
-        var recipeField = new RecipeFieldDomainEntity(fieldName, fieldData);
+        const string imageUrl = "http://wwww.google.com";
+        var recipeField = new AuthorDomainEntity(authorName, imageUrl);
 
         // Call
         int totalLength = recipeField.TotalLength;
 
         // Assert
-        int expectedLength = fieldName.Length + fieldData.Length;
-        Assert.Equal(expectedLength, totalLength);
+        Assert.Equal(authorName.Length, totalLength);
     }
 }
