+++ test/RecipeBot.Domain.TestUtils/RecipeDomainEntityTestFactory.cs
@@ -19,7 +19,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Common.Utils;
-using RecipeBot.Domain.Entities;
+using RecipeBot.Domain.Models;
 
 namespace RecipeBot.Domain.TestUtils;
